Add Jupyter notebook for quick demo
In the notebook `Quick_demo_WavePy.ipynb`, I demonstrate the basic
features of the WavePy package. You can run elastic seismic wavefield
simulations, select data windows and compute sensitivity kernels.

In this demo, a P-wave sensitivity kernel is computed for a point force
source in the x direction. It is plotted together with the outline of
the P-wave Fresnel zone.